# docs/swagger.yaml
basePath: /api
definitions:
...
  ChangePasswordRequest:
    properties:
      confirmNewPassword:
        description: Must be the same as the newPassword value.
        type: string
      currentPassword:
        type: string
      newPassword:
        description: Min 6, max 150 characters.
        type: string
    type: object
...
  ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
...
  LoginRequest:
    properties:
      email:
        description: Must be unique
        type: string
      password:
        description: Min 6, max 150 characters.
        type: string
    type: object
...
  RegisterRequest:
    properties:
      confirmPassword:
        type: string
      email:
        description: Must be unique
        type: string
      password:
        description: Min 6, max 150 characters.
        type: string
      username:
        description: Min 3, max 30 characters.
        type: string
    type: object
  ResetPasswordRequest:
    properties:
      confirmNewPassword:
        description: Must be the same as the password value.
        type: string
      newPassword:
        description: Min 6, max 150 characters.
        type: string
      token:
        description: The token the user got from the email.
        type: string
    type: object
...
  VerifiedWithTokenRequest:
    properties:
      token:
        description: The token the user got from the email.
        type: string
    type: object
  VerifyEmailRequest:
    properties:
      email:
        type: string
    type: object
  handler.editReq:
    properties:
      email:
        description: Must be unique
        type: string
      image:
        description: image/png or image/jpeg
        format: binary
        type: string
      username:
        description: Min 3, max 30 characters.
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/model.HttpError'
    type: object
  model.ErrorsResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/model.FieldError'
        type: array
    type: object
  model.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  model.HttpError:
    properties:
      message:
        type: string
      type:
        type: string
    type: object
...
  model.Success:
    properties:
      success:
        type: boolean
    type: object
  model.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      image:
        type: string
      isOnline:
        type: boolean
      isVerified:
        type: boolean
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:4000
info:
  contact:
    email: khoingoc456@gmail.com
    name: KhoiPN
  description: MurmurChatapp REST API Specs. This service uses sessions for authentication
  title: MurmurChatapp API
  version: "1.0"
paths:
  /account:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get Current User
      tags:
      - Account
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Update Account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/handler.editReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Update Current User
      tags:
      - Account
...
  /account/change-password:
    put:
      consumes:
      - application/json
      parameters:
      - description: Change Password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Change Current User's Password
      tags:
      - Account
  /account/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Forgot Password
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Forgot Password Request
      tags:
      - Account
  /account/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: User Login
      tags:
      - Account
  /account/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
      summary: User Logout
      tags:
      - Account
...
  /account/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Create an Account
      tags:
      - Account
  /account/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset Password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Reset Password
      tags:
      - Account
  /account/verified-with-token:
    post:
      consumes:
      - application/json
      parameters:
      - description: Verified With Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/VerifiedWithTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Verified With Token
      tags:
      - Account
  /account/verify-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: Verify Email
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Verify Email
      tags:
      - Account
...
swagger: "2.0"